refactoring of fallback inner class to usual class
